refactor(card): use render state info method
